backend/requirements.txt
google-cloud-video-live-stream
fastapi [all]
requests-oauth2client

